Rename Checkmark file into Shapes
